add first docs, fix readme
