+++ gulpfile.js
@@ -3,6 +3,6 @@ var jsdoc = require('gulp-jsdoc'),
     gulp = require('gulp');
 
 gulp.task('jsdoc', function () {
-  gulp.src('./src/**/*.js')
+  gulp.src(['./src/**/*.js', "readme.md"])
     .pipe(jsdoc('../javascript-algorithms-docs'));
 });
