+++ src/graphs/shortest-path/bellman-ford.js
@@ -7,7 +7,42 @@
     this.weight = weight;
   }
 
-  // Complexity O(|V||E|)
+  /**
+   * Bellman–Ford algorithm computes shortest paths from a single source
+   * vertex to all of the other vertices in a weighted digraph (negative weights allowed).<br><br>
+   * Time complexity: O(|V||E|) where V and E are the number of vertices and edges respectively.
+   *
+   * @public
+   * @module graphs/shortest-path/bellman-ford
+   * @param {Array} vertexes Vertices of the graph.
+   * @param {Array} edges Edges of the graph.
+   * @param {Number} source Start vertex.
+   * @returns {Object} Object with two arrays (parents and distances) with shortest-path information.
+   *
+   * @example
+   * require('path-to-algorithms/src/graphs/shortest-path/bellman-ford');
+   * 
+   * var glob = (typeof window === 'undefined') ? global : window;
+   * var Edge = glob.Edge;
+   * var bellmanFord = glob.bellmanFord;
+   * var edges = [];
+   * var vertexes = [0, 1, 2, 3, 4];
+   * 
+   * edges.push(new Edge(0, 1, -1));
+   * edges.push(new Edge(0, 2, 4));
+   * edges.push(new Edge(1, 2, 3));
+   * edges.push(new Edge(1, 3, 2));
+   * edges.push(new Edge(3, 1, 1));
+   * edges.push(new Edge(4, 3, -3));
+   * edges.push(new Edge(1, 4, 2));
+   * edges.push(new Edge(3, 2, 5));
+   *
+   * // { 
+   * //   parents:   { '0': null, '1':  0, '2': 1, '3':  4, '4': 1 },
+   * //   distances: { '0': 0,    '1': -1, '2': 2, '3': -2, '4': 1 }
+   * // }
+   * var pathInfo = bellmanFord(vertexes, edges, 0);
+   */
   function bellmanFord(vertexes, edges, source) {
     var distances = {}, parents = {}, c;
     for (var i = 0; i < vertexes.length; i += 1) {
@@ -39,19 +74,3 @@
   global.bellmanFord = bellmanFord;
 
 }((typeof window === 'undefined') ? global : window));
-
-// var glob = (typeof window === 'undefined') ? global : window;
-// var Edge = glob.Edge;
-// var bellmanFord = glob.bellmanFord;
-// var edges = [];
-// var vertexes = [0, 1, 2, 3, 4];
-// edges.push(new Edge(0, 1, -1));
-// edges.push(new Edge(0, 2, 4));
-// edges.push(new Edge(1, 2, 3));
-// edges.push(new Edge(1, 3, 2));
-// edges.push(new Edge(3, 1, 1));
-// edges.push(new Edge(4, 3, -3));
-// edges.push(new Edge(1, 4, 2));
-// edges.push(new Edge(3, 2, 5));
-// 
-// console.log(bellmanFord(vertexes, edges, 0));
